==> drug_treated_samples/00_process_dia.Rmd <==


```{r}
library(dplyr)
library(MSnSet.utils)
library(msigdbr)

source("helper_scripts.R")
source("synapseUtil.R")
source("plotting enrichment.R")

dcolors=c(none='#CCCAAA',Gilteritinib='#0072B2',
          `Venetoclax`='#EEDD55',
          `Decitabine`='#D14610',
          `Gilteritinib+Venetoclax`='#05820F',
          `Decitabine+Venetoclax`='#EDAE49',
          `Gilteritinib+Decitabine`='#D172B2',
          `Gilteritinib+Decitabine+Venetoclax`='#666666')

pcolors=c(`035`='#669999',`292`="#AABB66",`335`="#BBAABB",`507`="#CCBB33")


syn<-synapseLogin()
crosstab_prot <- read.table(syn$get('syn64604961')$path,sep='\t')#"crosstab_PTRC_wrkpckg5908_DIA_nolib.txt", sep = "\t")
# syn <- synapseLogin()
# crosstab_prot <- read.table(syn$get('syn64604961')$path)
# table(grepl(";", rownames(crosstab_prot)))

crosstab_prot <- crosstab_prot[!grepl(";", rownames(crosstab_prot)), ]
# write.table(rownames(crosstab_prot), quote = F, file = "dia_proteins.txt", col.names = F, row.names = F)
conv_df <- read.table("idmapping_2024_08_27.tsv", sep = "\t", header = T) %>%
   group_by(From) %>%
   mutate(total = n()) %>% ungroup() %>%
   group_by(To) %>% mutate(total2 = n()) %>% ungroup()

## This picks the 'canonical' or 'default' result, if many symbols map to the same uniprot.
conv_df_unique <- conv_df %>%
   group_by(From) %>% slice(1)

# A6NJR5 Q6ZSR9 C9JQL5 could not be mapped. We use merge to exclude this results from crosstab_df
crosstab_df <- crosstab_prot %>% as.data.frame() %>%
   mutate(From = rownames(crosstab_prot)) %>%
   merge(conv_df_unique, by = "From") %>%
   mutate(feature = To) %>% select(-From, -To, -total, -total2)

# which(table(crosstab_df$feature) == 2)
## Combining two rows with repeated feature name (MOCS2), 
## log transforming then taking the mean samplewise for MOCS2
crosstab_df <- crosstab_df %>%
   tidyr::pivot_longer(-feature, values_to = "value", names_to = "sample") %>%
   mutate(value = log2(value)) %>%
   tidyr::pivot_wider(names_from = "sample", values_from = "value", values_fn = mean) 
crosstab <- crosstab_df %>%
   select(-feature) %>% 
   as.matrix()
colnames(crosstab) <- sub("^X", "", colnames(crosstab))

rownames(crosstab) <- crosstab_df$feature
meta <- data.frame(sample = colnames(crosstab)) %>%
   mutate(Treatment = sub("^.*_", "", sample),
          patient = sub("_.*$", "", sample))

##add in full drug names
druglist=list(UT='none',D='Decitabine',V='Venetoclax',G='Gilteritinib',
              GV='Gilteritinib+Venetoclax',GD='Gilteritinib+Decitabine',DV='Decitabine+Venetoclax',
              GVD='Gilteritinib+Decitabine+Venetoclax')
meta<-meta|>mutate(Drug=sapply(Treatment,function(x) druglist[[x]]))

meta<-meta|>subset(Treatment%in%c("UT","V","DV","D"))

rownames(meta) <- meta$sample
m <- MSnSet(exprs = crosstab[,meta$sample], pData = meta)
# pData(m)
# exprs(m)
   

hist(rowSums(is.na(exprs(m))))
nrow(m)

m <- m[rowSums(is.na(exprs(m))) <= 16, ]
hist(rowSums(is.na(exprs(m))))
nrow(m)
col_medians <- apply(exprs(m), 2, median, na.rm = T)
exprs(m) <- sweep(exprs(m), 2, col_medians, FUN = '-')
 
plot_pca(m, phenotype = "patient")+scale_fill_manual(values=pcolors)+scale_color_manual(values=pcolors)

ggsave('patientPCA_uc.pdf')
plot_pca(m, phenotype = "Drug")+scale_fill_manual(values=dcolors)+scale_color_manual(values=dcolors)
ggsave('drugPCA_uc.pdf')

plot_pca(m, phenotype = "Drug",components=2:3)+scale_fill_manual(values=dcolors)+scale_color_manual(values=dcolors)
write.table(exprs(m), "ptrc_4patient_cellines_dia.txt", sep = "\t", quote = F)

saveRDS(m, "msnset_4patients_uncorrected.RDS")
synapseStore("msnset_4patients_uncorrected.RDS",'syn64134071')


```



## Remove effect of individual patients


```{r}
m_corrected <- correct_batch_effect_NA(m, "patient", par.prior = TRUE)
saveRDS(m_corrected, "msnset_4patients_corrected.RDS")
synapseStore("msnset_4patients_corrected.RDS",'syn64961545')
plot_pca(m_corrected, "Drug")+scale_fill_manual(values=dcolors)+scale_color_manual(values=dcolors)
ggsave('drugPCA_corrected.pdf')
plot_pca(m_corrected, "patient")+scale_fill_manual(values=pcolors)+scale_color_manual(values=pcolors)
ggsave("patientPCA_corrected.pdf")

write.table(exprs(m_corrected), "4patients_corrected_crosstab.tsv", sep = "\t", quote = F)
synapseStore("4patients_corrected_crosstab.tsv",'syn64961545')

nmmat<-exprs(m_corrected)
nmmat[which(is.na(nmmat),arr.ind=T)]<-0.0

library(umap)
res<-umap(t(nmmat))

pu<-res$layout
colnames(pu)<-c('Dim1','Dim2')
pu<-pu[rownames(meta),]
pu<-cbind(pu,meta)

ggplot(pu,aes(x=Dim1,y=Dim2,col=Drug))+geom_point()
```
