==> drug_treated_samples/01_decVenExVivoExpression.Rmd <==
---
title: "Leverage ex vivo data to identify patterns of Ven/Dec resistance"
author: "Sara gosline"
date: "2025-02-28"
output: html_document
---

Here we load the normalized ex vivo data to identify specific proteins that are up-regulated and down-regulated after 1-mont drug treatment

## Pull data

```{r}
library(dplyr)
library(MSnSet.utils)
library(msigdbr)

source("helper_scripts.R")
source("synapseUtil.R")
source("plotting enrichment.R")
source("ratio_based_functions.R")
syn<-synapseLogin()
## Load corrected exp25 dataset
# m_corrected <- readRDS("msnset_4patients_corrected.RDS")
m_corrected <- readRDS(syn$get("syn64886977")$path)
pcolors=c(`035`='#669999',`292`="#AABB66",`335`="#BBAABB",`507`="#CCBB33")


```


## Calculate diffex across treatments with Dec and Ven

We can calculate signatures across individual drugs and shared drugs. Here we focus only on decitabine and venetoclax

```{r diffex}

meta<-pData(m_corrected)
meta$hasVenetoclax=rep("NoVenetoclax",nrow(meta))
meta$hasVenetoclax[grep("V",meta$Treatment)]='Venetoclax'
meta$hasVenetoclax[grep("UT",meta$Treatment)]<-'Untreated'

meta$hasDecitabine=rep('NoDecitabine',nrow(meta))
meta$hasDecitabine[grep("D",meta$Treatment)]='Decitabine'
meta$hasDecitabine[grep("UT",meta$Treatment)]<-'Untreated'


meta$hasGilteritinib=rep('NoGilteritinib',nrow(meta))
meta$hasGilteritinib[grep("G",meta$Treatment)]='Gilteritinib'
meta$hasGilteritinib[grep("UT",meta$Treatment)]<-'Untreated'

pData(m_corrected)<-meta

treatment_contrasts <- c("D-UT", "G-UT", "V-UT", "GV-UT", "GD-UT", "DV-UT", "GVD-UT","GV-V","DV-V",'DV-D',"GVD-GV")

treatment_contrasts <- c("D-UT", "V-UT", "DV-UT", "DV-V",'DV-D')

treats=list(`D-UT`='Decitabine',`G-UT`='Gilteritinib', 
            `V-UT`='Venetoclax', `GV-UT`='Gilteritinib+Venetoclax', 
            `GD-UT`='Gilteritinib+Decitabine', 
            `DV-UT`='Decitabine+Venetoclax', 
            `GVD-UT`='Gilteritinib+Decitabine+Venetoclax')

indiv_diffexp<- diffexp_helper(m_corrected, "Treatment", treatment_contrasts)

pat_diffexp<-indiv_diffexp#rbind(ven_diffexp,gil_diffexp,dec_diffexp,indiv_diffexp)
#other_diffexp <- diffexp_helper(m_corrected, "Treatment", 
#                              c('GV-V','DV-V'))

write.table(pat_diffexp, "dec_ven_patient_lines_diffexp.txt", sep = "\t", quote = F, row.names = F)
synapseStore("dec_ven_patient_lines_diffexp.txt",parentId='syn64961545')
# pat_diffexp <- read.table("patient_lines_diffexp.txt", sep = "\t", header = T)
#pat_diffexp <- read.table(syn$get("syn64605136")$path, sep = "\t", header = T)|>
#  mutate(Treatment=sapply(contrast,function(x) treats[[x]]))

## For counts
table(pat_diffexp %>% filter(adj.P.Val < 0.05) %>% pull(contrast),
      pat_diffexp %>% filter(adj.P.Val < 0.05) %>% pull(logFC) > 0)


```




## Signature comparison across drug treatments
We can evaluate signatures by individual drug/combination treatments or by pooling them across samples.


```{r}
#treatment_contrasts <- c("Venetoclax-NoVenetoclax","Venetoclax-Untreated")
## Diffexp count barplot here!!
table(pat_diffexp$contrast, pat_diffexp$adj.P.Val < 0.05)

##add an additional LFC filter
table(pat_diffexp$contrast, pat_diffexp$adj.P.Val < 0.05&abs(pat_diffexp$logFC)>1)

diffEx_desc<-pat_diffexp|>
  mutate(direction=ifelse(logFC>0,'Upregulated','Downregulated'))|>
  mutate(sig=adj.P.Val<0.05)|>
  mutate(Twofold=abs(logFC)>1)

dcounts<-diffEx_desc|>
  group_by(contrast,sig,Twofold,direction)|>
  summarize(nFeatures=n())|>
  subset(sig)

dcounts[grep('UT',dcounts$contrast),]|>
  mutate(drug=sapply(contrast,function(x) treats[[x]]))|>
ggplot(aes(x=direction,fill=drug,y=nFeatures))+
  geom_bar(stat='identity',position='dodge')+
  facet_grid(Twofold~.)+ggtitle("Number of differentially expressed proteins of drug changes compared to untreated")+scale_fill_manual(values=dcolors)+coord_flip()


dcounts<-diffEx_desc|>
  group_by(contrast,sig,direction)|>
  summarize(nFeatures=n())|>
  subset(sig)

dcounts[grep('UT',dcounts$contrast),]|>
  mutate(drug=sapply(contrast,function(x) treats[[x]]))|>
ggplot(aes(x=direction,fill=drug,y=nFeatures))+
  geom_bar(stat='identity',position='dodge')+
  ggtitle("Number of differentially expressed proteins of drug changes compared to untreated")+scale_fill_manual(values=dcolors)+coord_flip()

ggsave('venDecDiffexCounts.pdf')

library(UpSetR)


###add some labels and stuff to break up our signature analysis...
sigs<-diffEx_desc|>
  subset(sig)|>
  dplyr::select(feature,contrast,direction,sig)|>
  distinct()|>
#  mutate(sigName=ifelse(Twofold,'Two-fold','Less than two-fold'))|>
  tidyr::unite(col='Sig',contrast,direction,remove=FALSE)|>
  tidyr::unite(col='Signature',contrast,remove=FALSE)
# 

ut_over=sigs[grep("UT",sigs$contrast),]|>
  #subset(Twofold)|>
    dplyr::select(feature,Sig)|>
  mutate(val=1)|>
  tidyr::pivot_wider(names_from=Sig,values_from=val,values_fill = 0)|>
  tibble::column_to_rownames('feature')|>
  upset(nsets=16,nintersects=40,text.scale=2)
ut_over

```



## Visualizing the protins across the patient samples

Now let's look at the effects of the different features across the samples


```{r signature heatmaps}


f=subset(sigs,Signature=='V-UT')

vprots=f$feature

nog_pats<-rownames(subset(pData(m_corrected),Treatment%in%c('UT','V','D',"DV")))

pData(m_corrected)<-as.data.frame(apply(pData(m_corrected),2,as.factor))|>
    mutate(Drug=sapply(Treatment,function(x) druglist[[x]]))

plot_heatmap(m_corrected[,nog_pats], features=vprots, annotation_cols=c("Drug",'patient'), 
             ann_colors = list(Drug=dcolors,patient=pcolors),cluster_cols=TRUE,clustering_method='ward.D2',
             #,patient=pcolors), 
             heatmap_title = "Ven signature",file='venUTHeamap.pdf')


##now we can visualize in boxplots as well
prot_dat<-exprs(m_corrected[,nog_pats])|>
  as.data.frame()|>
  tibble::rownames_to_column('feature')|>
  tidyr::pivot_longer(cols=contains('_'),
                      names_to='sample',values_to='logExpr')|>
  left_join(pData(m_corrected))|>
  left_join(select(diffEx_desc,c(feature,contrast,direction,sig)))

prot_dat$Drug=factor(prot_dat$Drug,
                     levels=c('none','Decitabine','Venetoclax','Decitabine+Venetoclax'))


p<-prot_dat|>
  subset(!is.na(direction))|>
  subset(feature%in%vprots)|>#sig)|>subset(Twofold)|>
  ggplot(aes(x=direction,y=logExpr,fill=Drug))+
  geom_boxplot()+facet_grid(~patient)+scale_fill_manual(values=dcolors)+
  ggtitle("Ven sig across patients")

p

ggsave('venSignatureAcrossPatients.pdf')

###now lets move to doublec ombinations


f=subset(sigs,Signature=='DV-UT')
dvprots=f$feature

plot_heatmap(m_corrected[,nog_pats], features=dvprots, annotation_cols=c("Drug",'patient'), 
             ann_colors = list(Drug=dcolors,patient=pcolors),cluster_cols=TRUE,
             heatmap_title = "DV-UT",clustering_method='ward.D2',file='decVen_UTHeamap.pdf')



p<-prot_dat|>
  subset(!is.na(direction))|>
  subset(feature%in%dvprots)|>#sig)|>subset(Twofold)|>
  ggplot(aes(x=direction,y=logExpr,fill=Drug))+
  geom_boxplot()+facet_grid(~patient)+scale_fill_manual(values=dcolors)+
  ggtitle("DV sig across patients")

p

ggsave('decVenSignatureAcrossPatients.pdf')
# 
# f=subset(sigs,Signature=='GV-UT_Two-fold')
# gvprots=f$feature
# 
# plot_heatmap(m_corrected, features=gvprots, annotation_cols=c("Drug",'patient'), 
#              ann_colors = list(Drug=dcolors,patient=pcolors),cluster_cols=FALSE, 
#              heatmap_title = "GV-UT_Two-fold",file='giltVen_UTHeatmap.pdf')
# 
# p<-prot_dat|>
#   subset(!is.na(direction))|>
#   subset(feature%in%gvprots)|>#sig)|>subset(Twofold)|>
#   ggplot(aes(x=direction,y=logExpr,fill=Drug))+
#   geom_boxplot()+facet_grid(~patient)+scale_fill_manual(values=dcolors)+
#   ggtitle("GV sig across patients")
# 
# p
# ggsave('giltVenSignatureAcrossPatients.pdf')
# 
# 
# 
# f=subset(sigs,Signature=="GVD-UT_Two-fold")
# gvdprots<-f$feature
# 
# 
# plot_heatmap(m_corrected, features=gvdprots, annotation_cols=c("Drug",'patient'), 
#              ann_colors = list(Drug=dcolors,patient=pcolors),#,patient=pcolors), 
#              heatmap_title = "GVD-UT_Two-fold",cluster_cols=FALSE,
#              file='gvd_UTHeatmap.pdf')
# 
# 
# p<-prot_dat|>
#   subset(!is.na(direction))|>
#   subset(feature%in%gvdprots)|>#sig)|>subset(Twofold)|>
#   ggplot(aes(x=direction,y=logExpr,fill=Drug))+
#   geom_boxplot()+facet_grid(~patient)+scale_fill_manual(values=dcolors)+ggtitle('GVD sig acrosspatients')
# 
# p
# ggsave('decVenGiltSignatureAcrossPatients.pdf')
# 
# print(paste('There are',length(intersect(gvprots,gvdprots)),'features in common between gvd and gv'))
# 
# gvonly<-setdiff(gvprots,gvdprots)
# 
# plot_heatmap(m_corrected, features=gvonly, annotation_cols=c("Drug",'patient'), 
#              ann_colors = list(Drug=dcolors,patient=pcolors), 
#              heatmap_title = "GV-only",cluster_cols=FALSE,
#              file='gvOnly_UTHeatmap.pdf')
# 
# p<-prot_dat|>
#   subset(!is.na(direction))|>
#   subset(feature%in%gvonly)|>#sig)|>subset(Twofold)|>
#   ggplot(aes(x=direction,y=logExpr,fill=Drug))+
#   geom_boxplot()+facet_grid(~patient)+scale_fill_manual(values=dcolors)+ggtitle('GV only sig acrosspatients')
# 
# p
# ggsave('venGiltOnlySignatureAcrossPatients.pdf')
# 
# gvdonly<-setdiff(gvdprots,gvprots)
# 
# 
# plot_heatmap(m_corrected, features=gvdonly, annotation_cols=c("Drug",'patient'), 
#              ann_colors = list(Drug=dcolors,patient=pcolors), 
#              heatmap_title = "GVD-only",cluster_cols=FALSE,file='gvdOnly_UTheatmap.pdf')
# 
# p<-prot_dat|>
#   subset(!is.na(direction))|>
#   subset(feature%in%gvdonly)|>#sig)|>subset(Twofold)|>
#   ggplot(aes(x=direction,y=logExpr,fill=Drug))+
#   geom_boxplot()+facet_grid(~patient)+scale_fill_manual(values=dcolors)+ggtitle('GVD only sig acrosspatients')
# 
# p
# ggsave('venGiltDecOnlySignatureAcrossPatients.pdf')
# 
# ##what if we subtract out other proteins

```


## GSEA enrichment


First we get all the gene signatures from MSIGDGB

```{r}
t2g_hallmark <- msigdbr(species = "Homo sapiens", collection = "H") %>%
  dplyr::select(gs_name, gene_symbol)

t2g_gobp <- msigdbr(species = "Homo sapiens", collection = "C5", subcollection = "BP") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))

t2g_gocc <- msigdbr(species = "Homo sapiens", collection = "C5", subcollection = "CC") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))

t2g_gomf <- msigdbr(species = "Homo sapiens", collection = "C5", subcollection = "MF") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))

t2g_wikipath <- msigdbr(species = "Homo sapiens", collection = "C2", subcollection = "WIKIPATHWAYS") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))

t2g_reactome <- msigdbr(species = "Homo sapiens", collection = "C2", subcollection = "REACTOME") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))


t2g_kegg <- msigdbr(species = "Homo sapiens", collection = "C2", subcollection = "KEGG_MEDICUS") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))


all_t2gs <- list(t2g_gobp, t2g_gocc, t2g_gomf, t2g_hallmark, t2g_kegg, t2g_reactome)
names(all_t2gs) <- c("GOBP", "GOCC", "GOMF", "Hallmark", "KEGG", "Reactome")

```

Then we evaluate everything for each contrats. 

```{r, message=FALSE, warning=FALSE,echo=FALSE}
## Getting the GSEA results

#pData(m_corrected)=pData(m_corrected)[,c('sample','Treatment','patient','Drug')]
 for (t2g_name in names(all_t2gs)){
   GSEA_helper(m_corrected, treatment_contrasts, t2g = all_t2gs[[t2g_name]],
               t2g_name = paste0("ptrc_4patient_lines_", t2g_name), "Treatment")
 }


 GSEA_global <- rbind(read.table("GSEA_ptrc_4patient_lines_GOBP_combined.txt", sep = "\t") %>% 
                        mutate(DB = "GOBP"),
                      read.table("GSEA_ptrc_4patient_lines_GOCC_combined.txt", sep = "\t") %>% 
                        mutate(DB = "GOCC"),
                      read.table("GSEA_ptrc_4patient_lines_GOMF_combined.txt", sep = "\t") %>% 
                        mutate(DB = "GOMF"),
                      read.table("GSEA_ptrc_4patient_lines_Hallmark_combined.txt", sep = "\t") %>% 
                        mutate(DB = "Hallmark"),
                      read.table("GSEA_ptrc_4patient_lines_KEGG_combined.txt", sep = "\t") %>% 
                        mutate(DB = "KEGG"),
                      read.table("GSEA_ptrc_4patient_lines_Reactome_combined.txt", sep = "\t") %>% 
                        mutate(DB = "Reactome"))
 write.table(GSEA_global, "patient_lines_GSEA.txt", sep = "\t", quote = F, row.names = F)

# GSEA_global <- read.table("GSEA tables/patient_lines_GSEA.txt", sep = "\t", header = T)
#GSEA_global <- read.table(syn$get("syn64605143")$path, sep = "\t", header = T)
# GSEA_210_ven <- read.table("GSEA tables/GSEA_ven_resistant_vs_sensitive_210cohort.txt", sep = "\t", header = T)
#GSEA_210_ven <- read.table(syn$get("syn64605142")$path, sep = "\t", header = T)

```




Plotting the GSEA results



```{r,warning=FALSE,message=FALSE}
enrichment_df <- GSEA_global %>%
   filter(contrast == "V-UT",
          DB == "Hallmark") %>% 
   mutate(pathway = sub("HALLMARK_", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = NES,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "V-UT", enrichment_title = "Hallmark GSEA Venetoclax - Untreated")

enrichment_df <- GSEA_global %>%
   filter(contrast == "DV-UT",
          DB == "Hallmark") %>% 
   mutate(pathway = sub("HALLMARK_", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = NES,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "DV-UT", enrichment_title = "Hallmark GSEA DV - UT")

enrichment_df <- GSEA_global %>%
   filter(contrast == "GVD-UT",
          DB == "Hallmark") %>% 
   mutate(pathway = sub("HALLMARK_", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = NES,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "GVD-UT", enrichment_title = "Hallmark GSEA GVD - UT")


enrichment_df <- GSEA_global %>%
   filter(contrast == "V-UT",
          DB == "Reactome") %>% 
   mutate(pathway = sub("REACTOME ", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = NES,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "V-UT", enrichment_title = "Reactome GSEA Venetoclax -  Untreated")


enrichment_df <- GSEA_global %>%
   filter(contrast == "DV-UT",
          DB == "Reactome") %>% 
   mutate(pathway = sub("REACTOME", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = NES,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "DV-UT", enrichment_title = "Hallmark GSEA DV - UT")

enrichment_df <- GSEA_global %>%
   filter(contrast == "GVD-UT",
          DB == "Reactome") %>% 
   mutate(pathway = sub("REACTOME", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = NES,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "GVD-UT", enrichment_title = "Hallmark GSEA GVD - UT")


 subset(GSEA_global,DB%in%c("KEGG"))|>subset(contrast%in%c('D-UT','DV-UT','DV-V','DV-D','V-UT'))|>subset(p.adjust<0.05)|>ggplot(aes(x=contrast,y=Description,col=NES,size=(-1*log10(p.adjust))))+geom_point()+scale_color_gradient2(low='slateblue',mid='darkgrey',high='darkred')

  ggsave('keggResults.pdf')

 subset(GSEA_global,DB%in%c("Reactome"))|>subset(contrast%in%c('D-UT','DV-UT','DV-V','DV-D','V-UT'))|>subset(p.adjust<0.01)|>ggplot(aes(x=contrast,y=Description,col=NES,size=(-1*log10(p.adjust))))+geom_point()+scale_color_gradient2(low='slateblue',mid='darkgrey',high='darkred')
 
 ggsave('ReactomeResults.pdf',height=12,width=14)

 subset(GSEA_global,DB%in%c("Hallmark"))|>subset(contrast%in%c('D-UT','DV-UT','DV-V','DV-D','V-UT'))|>subset(p.adjust<0.05)|>ggplot(aes(x=contrast,y=Description,col=NES,size=(-1*log10(p.adjust))))+geom_point()+scale_color_gradient2(low='slateblue',mid='darkgrey',high='darkred')
 
 ggsave('hallmarkResults.pdf')
```
### Focus on IFNgamma signal

IFNGamma has been seen as a signal that improves prognosis, generally communicated by T-cells and NK cells. Is there expression in these cells?

```{r ifng}

##first lets only focus on samples without gilteritinib


ifngs<-unlist(strsplit(subset(GSEA_global,Description=='HALLMARK_INTERFERON_GAMMA_RESPONSE')|>subset(contrast=='DV-UT')|>pull('core_enrichment'),split='/'))

#nmiss<-complete.cases(exprs(m_corrected))

plot_heatmap(m_corrected[,nog_pats], features=ifngs, annotation_cols=c("Drug",'patient'), 
             ann_colors = list(Drug=dcolors,patient=pcolors), 
             heatmap_title = "IFNG proteins",cluster_cols=TRUE,file='ifnGheatmap.pdf')

ifnas<-unlist(strsplit(subset(GSEA_global,Description=='HALLMARK_INTERFERON_ALPHA_RESPONSE')|>subset(contrast=='DV-UT')|>pull('core_enrichment'),split='/'))

#nmiss<-complete.cases(exprs(m_corrected))

plot_heatmap(m_corrected[,nog_pats], features=ifnas, annotation_cols=c("Drug",'patient'), 
             ann_colors = list(Drug=dcolors,patient=pcolors), 
             heatmap_title = "IFNa proteins",cluster_cols=TRUE,file='ifnAheatmap.pdf')



mycv2<-unlist(strsplit(subset(GSEA_global,Description=='HALLMARK_MYC_TARGETS_V2')|>subset(contrast=='DV-UT')|>pull('core_enrichment'),split='/'))

plot_heatmap(m_corrected[,nog_pats], features=mycv2, annotation_cols=c("Drug",'patient'), 
             ann_colors = list(Drug=dcolors,patient=pcolors), 
             heatmap_title = "MYCV2 proteins",cluster_cols=TRUE)

mycv1<-unlist(strsplit(subset(GSEA_global,Description=='HALLMARK_MYC_TARGETS_V1')|>subset(contrast=='DV-UT')|>pull('core_enrichment'),split='/'))

plot_heatmap(m_corrected[,nog_pats], features=mycv1, annotation_cols=c("Drug",'patient'), 
             ann_colors = list(Drug=dcolors,patient=pcolors), 
             heatmap_title = "MYCV1 proteins",cluster_cols=TRUE)
```



### Over representation of differential expression
Here we can use the basic signatures.

```{r, message=FALSE,warning=FALSE, error=FALSE}

## Using ex vivo

diffexp_ora <- pat_diffexp %>% mutate(adj_pval = adj.P.Val) %>%
   select(feature, logFC, adj_pval, contrast)

# gvonly_ora<-subset(diffexp_ora,contrast=="GV-UT")|>
#   mutate(contrast='GV Only')|>
#   subset(feature%in%gvonly)
# 
# gvdonly_ora<-subset(diffexp_ora,contrast=="GVD-UT")|>
#     mutate(contrast='GVD Only')|>
#   subset(feature%in%gvdonly)
# 
# diffexp_ora<-rbind(diffexp_ora,gvonly_ora,gvdonly_ora)


ora_results_global <- rbind(ORA_helper(diffexp_ora, diffexp_ora$feature, t2g_gobp,logFC_cutoff=1) %>% 
                        mutate(DB = "GOBP"),
                     ORA_helper(diffexp_ora, diffexp_ora$feature, t2g_gocc,logFC_cutoff=0) %>% 
                        mutate(DB = "GOCC"),
                     ORA_helper(diffexp_ora, diffexp_ora$feature, t2g_gomf,logFC_cutoff=0) %>% 
                        mutate(DB = "GOMF"),
                     ORA_helper(diffexp_ora, diffexp_ora$feature, t2g_hallmark,logFC_cutoff=0) %>% 
                        mutate(DB = "Hallmark"),
                     ORA_helper(diffexp_ora, diffexp_ora$feature, t2g_kegg,logFC_cutoff=0) %>% 
                        mutate(DB = "KEGG"),
                     ORA_helper(diffexp_ora, diffexp_ora$feature, t2g_reactome,logFC_cutoff=0) %>% 
                        mutate(DB = "Reactome"),
                     ORA_helper(diffexp_ora, diffexp_ora$feature, t2g_wikipath,logFC_cutoff=0) %>% 
                        mutate(DB = "Wikipath"))


stats=ora_results_global|>subset(p.adjust<0.1)|>group_by(DB,contrast,sign)|>summarize(count=n())
print(stats)

stats|>ggplot(aes(x=contrast,y=count,fill=sign))+facet_grid(~DB)+geom_bar(stat='identity',position='dodge')+coord_flip()
```

Now we can plot all the conditions of interest

```{r plot ORA, message=FALSE,warning=FALSE}
enrichment_df <- ora_results_global %>%
   filter(contrast == "V-UT",
          DB == "Hallmark") %>% 
   mutate(pathway = sub("HALLMARK_", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = zScore,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "V-UT", enrichment_title = "Hallmark ORA Venetoclax - Untreated")


enrichment_df <- ora_results_global %>%
   filter(contrast == "DV-UT",
          DB == "Hallmark") %>% 
   mutate(pathway = sub("HALLMARK_", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = zScore,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)

plot_enrichment_result(enrichment_df, "DV-UT", enrichment_title = "Hallmark ORA DV - UT")

enrichment_df <- ora_results_global %>%
   filter(contrast == "GVD-UT",
          DB == "Hallmark") %>% 
   mutate(pathway = sub("HALLMARK_", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = zScore,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "GVD-UT", enrichment_title = "Hallmark ORA GVD - UT")


enrichment_df <- ora_results_global %>%
   filter(contrast == "V-UT",
          DB == "Reactome") %>% 
   mutate(pathway = sub("REACTOME ", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = zScore,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "V-UT", enrichment_title = "Reactome ORA Venetoclax -  Untreated")


enrichment_df <- ora_results_global %>%
   filter(contrast == "DV-UT",
          DB == "Reactome") %>% 
   mutate(pathway = sub("REACTOME", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = zScore,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "DV-UT", enrichment_title = "Hallmark ORA DV - UT")

enrichment_df <- ora_results_global %>%
   filter(contrast == "GVD-UT",
          DB == "Reactome") %>% 
   mutate(pathway = sub("REACTOME", "", Description),
          pathway = gsub("_", " ", pathway),
          enrichment = zScore,
          adj_p_val = p.adjust) %>%
   arrange(adj_p_val)
plot_enrichment_result(enrichment_df, "GVD-UT", enrichment_title = "Hallmark ORA GVD - UT")
```


## Trying out networks
Since the number of diffex genes is so small, maybe we can build networks to visualize activity.


```{r}
library(PCSF)
data('STRINGv12')

ppi <- construct_interactome(STRINGv12)

#subnet <- PCSF_rand(ppi,abs(terms), n=100, r=0.2,w = 2, b = 2, mu = 0.0005,...)

#######################################
nets<-lapply(c("DV-UT", "DV-V","V-UT","D-UT"),function(con){
  ups<-subset(diffEx_desc,direction=='Upregulated')|>subset(contrast==con)|>subset(adj.P.Val<0.1)
  upnet<-NULL
  if(nrow(ups)>2){
    uprots<-ups$logFC
    names(uprots)<-ups$feature
    try(upnet<-PCSF_rand(ppi,abs(uprots), n=100, r=0.2,w = 2, b = 1, mu = 0.0005))
  }
  downet<-NULL
  downs<-subset(diffEx_desc,direction=='Downregulated')|>subset(contrast==con)|>subset(adj.P.Val<0.1)
  if(nrow(downs)){
        downprots<-downs$logFC
    names(downprots)<-downs$feature
    try(downet<-PCSF_rand(ppi,abs(downprots), n=100, r=0.2,w = 2, b = 1, mu = 0.0005))
  }
  return(list(up=upnet,down=downet))
})

names(nets)<-c('DecVen_None','DecVen_Ven','Ven_None','Dec_None')
plot.PCSF(nets[[1]]$up)
plot.PCSF(nets[[1]]$down)

## V-UT compared to Ven Resistant vs Sensitive in 210 cohort.
#ven_res_vs_sen_pathway <- ora_cohort_venetoclax %>% filter(contrast == "Resistant-Sensitive", p.adjust < 0.05) 
#View(ora_results_global %>% filter(ID %in% ven_res_vs_sen_pathway$ID, contrast == "V-UT"))

#V_UT_pathways <- ora_results_global %>% filter(contrast == "V-UT", p.adjust < 0.05) 
#View(ora_cohort_venetoclax %>% filter(ID %in% V_UT_pathways$ID, contrast == "Resistant-Sensitive"))

```
