==> drug_treated_samples/01a_ratioOptimization.Rmd <==
---
title: "Leverage ex vivo data to identify signatures"
author: "Sara gosline"
date: "2025-02-28"
output: html_document
---

Here we load the normalized ex vivo data to identify specific proteins that are up-regulated and down-regulated after 1-mont drug treatment

## Pull data

```{r}
library(dplyr)
library(MSnSet.utils)
library(msigdbr)

source("helper_scripts.R")
source("synapseUtil.R")

syn = synapseLogin()
source("plotting enrichment.R")
source("ratio_based_functions.R")
pcolors=c(`035`='#669999',`292`="#AABB66",`335`="#BBAABB",`507`="#CCBB33")


## Load corrected exp25 dataset
# m_corrected <- readRDS("msnset_4patients_corrected.RDS")
m_corrected <- readRDS(syn$get("syn64886977")$path)

```


## Create signatures via different means

Determine signatures in different ways - lfc, p-value, combination, etc. 

```{r diffex}

meta<-pData(m_corrected)

treatment_contrasts <- c("D-UT", "V-UT",  "DV-UT")#,"GV-V","DV-V","GVD-GV")"G-UT","GV-UT", "GD-UT",, "GVD-UT"
treats=list(`D-UT`='Decitabine',#`G-UT`='Gilteritinib', 
            `V-UT`='Venetoclax', #`GV-UT`='Gilteritinib+Venetoclax', 
            #`GD-UT`='Gilteritinib+Decitabine', 
            `DV-UT`='Decitabine+Venetoclax')#, 
            #`GVD-UT`='Gilteritinib+Decitabine+Venetoclax')

meta<-subset(meta,Treatment%in%c('UT','D','V','DV'))


```




## Tweaking the signatures


I'm not sure what the best way is to tweak the signatures, but here are a series of tests we can perform.Right now we're just altering the LFC threshold

### Generate signatures

Choose a set of parameters and get signatures. 

```{r optimize}

lfc=0.5

pdf(paste0('signaturePlotsLFC',lfc,'.pdf'),width=11)
finalsigs<-do.call(rbind,lapply(treatment_contrasts,function(y){
    markers=getSigsFromData(condition=y,doPlot=TRUE,byPval=TRUE,lfcthresh=lfc)
    rbind(data.frame(treatment=treats[[y]],markers=markers$resistant,condition='resistant'),
          data.frame(treatment=treats[[y]],markers=markers$sensitive,condition='sensitive'))}))

dev.off()


##plotSigsOn original data
pdf(paste0('signaturePatientHeatmapsLFC',lfc,'.pdf'))
res<-lapply(unique(finalsigs$treatment),function(treat){
 # print(treat)
    cmeta<-meta[c('patient','Treatment')]
    cmeta$Treatment=sapply(cmeta$Treatment,function(x) ifelse(x=='UT','none',treats[[paste0(x,'-UT')]]))
     dat<-subset(finalsigs,treatment==treat)|>
      tibble::remove_rownames()
    expr<-exprs(m_corrected[dat$marker,rownames(subset(cmeta,Treatment%in%c('none',treat)))])
    expr[which(is.na(expr),arr.ind=TRUE)]<-0.0
   
    rmdata=tibble::column_to_rownames(dat,'markers')|>select(condition)|>mutate(condition=paste(condition,'signal'))
    #print(cmeta)
    #print(rmdata)
  pheatmap::pheatmap(expr,
                     annotation_col=cmeta,cluster_rows=FALSE,clustering_distance_cols = 'correlation',
                     annotation_colors = list(condition=sigcolors,Treatment=dcolors[unique(cmeta$Treatment)],patient=pcolors),cellwidth=10,
                     annotation_row=data.frame(rmdata),cellheight=10,main=paste(treat,'treated signature\n in patient samples'))
  
})
dev.off()
```

### Evaluate Concordance with per-patient scores

How does each patient sample score? We want to make sure the signatures themselves are concordant across all 4 patients. 

```{r}


res1<-do.call(rbind,lapply(treats,function(x){
   rtab<-subset(finalsigs,treatment==x)|>
     subset(condition=='resistant')
  stab<-subset(finalsigs,treatment==x)|>
     subset(condition=='sensitive')
  rat=calcRatio(rtab$markers,stab$markers,exprs(m_corrected))|>as.data.frame()|>
    tibble::rownames_to_column('sample')|>
    mutate(signature=x)
  rat
}))

res<-res1|>as_tibble()|>tidyr::separate_wider_delim(cols='sample',names=c('patient','treatment'),delim='_')|>
  subset(treatment%in%c('UT','D','V','DV'))
res$treatment=unlist(lapply(res$treatment,function(x) ifelse(x=='UT','none',treats[[paste0(x,'-UT')]])))

ggplot(res,aes(x=treatment,y=ratio,fill=signature))+geom_boxplot()+scale_fill_manual(values=dcolors)+coord_flip()


ggplot(res,aes(x=signature,y=ratio,fill=treatment))+geom_boxplot()+scale_fill_manual(values=dcolors)+coord_flip()

#ggplot(res,aes(y=signature,x=ratio,fill=treatment,alpha=0.8))+ggridges::geom_density_ridges()+scale_fill_manual(values=dcolors)

ggsave(paste0('signatureLFC',lfc,'ScoreOfSamples.pdf'),height=4)

```

### Calculate signature overlap
some of the signatures have a lot in common. how can we check?

```{r sign overlap}

library(UpSetR)
 sigprots<-finalsigs|>tidyr::pivot_wider(names_from=condition,values_from=markers)|>
   tidyr::unnest(cols=c(resistant,sensitive))
 
 
 resmat<-sigprots|>dplyr::select(treatment,resistant)|>
   mutate(value=1)|>
   tidyr::pivot_wider(names_from=treatment,values_from=value,values_fill=0)|>
   as.data.frame()#|>
   #tibble::column_to_rownames('signature')|>as.matrix()
 
 pdf(paste0('signatureLFC',lfc,'Overlap.pdf'))
 UpSetR::upset(resmat,nsets=7,main.bar.color=sigcolors[2])


 sensmat<-sigprots|>dplyr::select(treatment,sensitive)|>
   mutate(value=1)|>
   tidyr::pivot_wider(names_from=treatment,values_from=value,values_fill=0)|>
   as.data.frame()#|>
   #tibble::column_to_rownames('signature')|>as.matrix()
 
 UpSetR::upset(sensmat,nsets=7,main.bar.color=sigcolors[3])
 dev.off()
```
It looks like there are a few pairs of genes that show up in multiple signatures...
```{r dupes}

doub_sens=sigprots|>group_by(sensitive)|>summarize(numsigs=n_distinct(treatment))|>subset(numsigs>1)
doub_res=sigprots|>group_by(resistant)|>summarize(numsigs=n_distinct(treatment))|>subset(numsigs>1)

doub_res|>left_join(sigprots)|>dplyr::select(resistant,numsigs,treatment)
doub_sens|>left_join(sigprots)|>dplyr::select(sensitive,numsigs,treatment)

```
### Calculate Pathway enrichment

For each signature/paramtere we need to evaluate the functional enrichment of each signature to show what it is predicting. 

```{r go enrichment}
##first generate a list of libraries
t2g_hallmark <- msigdbr(species = "Homo sapiens", collection = "H") %>%
  dplyr::select(gs_name, gene_symbol)

t2g_gobp <- msigdbr(species = "Homo sapiens", collection = "C5", subcollection = "BP") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))

t2g_gocc <- msigdbr(species = "Homo sapiens", collection = "C5", subcollection = "CC") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))

t2g_gomf <- msigdbr(species = "Homo sapiens", collection = "C5", subcollection = "MF") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))

t2g_wikipath <- msigdbr(species = "Homo sapiens", collection = "C2", subcollection = "WIKIPATHWAYS") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))

t2g_reactome <- msigdbr(species = "Homo sapiens", collection = "C2", subcollection = "REACTOME") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))


t2g_kegg <- msigdbr(species = "Homo sapiens", collection = "C2", subcollection = "KEGG_MEDICUS") %>% 
  dplyr::select(gs_name, gene_symbol) %>%
  mutate(gs_name = gsub("_", " ", gs_name))


all_t2gs <- list(t2g_gobp, t2g_gocc, t2g_gomf, t2g_hallmark, t2g_kegg, t2g_reactome)
names(all_t2gs) <- c("GOBP", "GOCC", "GOMF", "Hallmark", "KEGG", "Reactome")

##now iterate through every signature and get terms enriched for sensitive and resistnat
##markers
universe=unique(sigs$feature)

eres<-do.call(rbind,lapply(names(all_t2gs),function(x){
  t2g=all_t2gs[[x]]
  do.call(rbind,lapply(treatment_contrasts,function(y){
    markers=getSigsFromData(condition=y,lfcthresh=lfc)
    #print(markers)
  res <- try(clusterProfiler::enricher(markers$resistant, 
                                               universe = universe, TERM2GENE = t2g,
                                               maxGSSize = 500, pvalueCutoff = 0.05)@result %>%
        mutate(sign = 'resistant', signature = treats[[y]],path=x))
  sens<-try(clusterProfiler::enricher(markers$sensitive, 
                                             universe = universe, TERM2GENE = t2g,
                                             maxGSSize = 500, pvalueCutoff = 0.05)@result %>%
      mutate(sign = 'sensitive', signature = treats[[y]],path=x))
  return(rbind(res,sens))
  }))
}))

eres$qvalue<-as.numeric(eres$qvalue)
eres$Count<-as.numeric(eres$Count)
eres|>subset(qvalue<0.25)|>subset(Count>1)|>group_by(sign,signature,path)|>summarize(nterms=n())|>ggplot(aes(x=path,y=nterms,fill=signature))+geom_bar(stat='identity',position='dodge')+facet_grid(sign~.)+scale_fill_manual(values=dcolors)

ggsave(paste0('sigLFC',lfc,'enrichedTermsQval25AtLeast2Proteins.pdf'))

readr::write_csv(eres,file=paste0('sigLFC',lfc,'functionalEnrichmentResults.csv'))
```
What terms are shared? are there anything interesting?/

```{r enrich results}

paths<-eres|>subset(qvalue<0.15)|>subset(as.numeric(Count)>1)|>
  subset(path%in%c('Reactome','GOBP','Hallmark','KEGG'))|>
  group_by(ID)|>summarize(nsigs=n_distinct(signature))#|>subset(nsigs>1)

eres$signature=factor(eres$signature,levels=treats)#c('Gilteritinib','Venetoclax','Decitabine','Decitabine+Venetoclax','Gilteritinib+Venetoclax','Gilteritinib+Decitabine','Gilteritinib+Decitabine+Venetoclax'))

##now create a bubble plot of the pathways by enrichment? 
eres|>subset(sign=='resistant')|>
  subset(ID%in%paths$ID)|>ggplot(aes(x=signature,y=ID,size=Count,col=qvalue))+geom_point()+ggtitle("Resistant signature enrichment")+theme(axis.text.x=element_text(angle=90, hjust=1))
ggsave(paste0('sigLFC',lfc,'resEnrichGOBPReactome.pdf'),height=10)

eres|>subset(sign=='sensitive')|>
  subset(ID%in%paths$ID)|>ggplot(aes(x=signature,y=ID,size=Count,col=qvalue))+geom_point()+ggtitle("Sensitive signature enrichment")+theme(axis.text.x=element_text(angle=90, hjust=1))
ggsave(paste0('sigLFC',lfc,'sensEnrichGOBPReactome.pdf'),height=12,width=12)

# 
# ##should we do reactome too?
# paths<-eres|>subset(qvalue<0.25)|>subset(Count>1)|>
#   subset(path=='Reactome')|>
#   group_by(ID)|>summarize(nsigs=n_distinct(signature))|>subset(nsigs>1)
# 
# ##now create a bubble plot of the pathways by enrichment? 
# eres|>subset(sign=='resistant')|>
#   subset(ID%in%paths$ID)|>ggplot(aes(x=signature,y=ID,size=Count,col=qvalue))+geom_point()+ggtitle("Resistant signature enrichment - Reactome")+theme(axis.text.x=element_text(angle=90, hjust=1))
# ggsave('resEnrichReactome.pdf',height=4)
# 
# eres|>subset(sign=='sensitive')|>
#   subset(ID%in%paths$ID)|>ggplot(aes(x=signature,y=ID,size=Count,col=qvalue))+geom_point()+ggtitle("Sensitive signature enrichment - Reactome")+theme(axis.text.x=element_text(angle=90, hjust=1))
# dev.off()
# ggsave('sensEnrichReactome.pdf',height=4)
```

## Other metrics?
How can we assess these things? can we look at publicly available cell lines
